refactor: 變更global.js的檔案位置
