+++ src/globalMethods/global.js
